scripts/social_media/pyproject.toml
[tool.poetry]
name = "social_media"
version = "0.1.0"
description = ""
authors = ["Global.health team <info@global.health>"]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
iso3166 = "^2.1.1"
requests = "^2.28.1"
pygsheets = "^2.0.5"
tweepy = "^4.10.1"
emoji-country-flag = "^1.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
